ATO-1022: moves validateAuthorisationHeader into own function
