+++ backend/api/src/handlers/dynamodb/get-service-clients.ts
@@ -1,5 +1,6 @@
 import {APIGatewayProxyEvent, APIGatewayProxyResult} from "aws-lambda";
 import DynamoDbClient from "../../dynamodb-client";
+import {validateAuthorisationHeader} from "../helper/validate-authorisation-header";
 
 const client = new DynamoDbClient();
 
@@ -12,7 +13,7 @@ export const getServiceClientsHandler = async (event: APIGatewayProxyEvent): Pro
         };
     }
 
-    const authHeaderValidationResult = validateAuthorizationHeader(event.headers.Authorization);
+    const authHeaderValidationResult = validateAuthorisationHeader(event.headers.Authorization);
 
     if (!authHeaderValidationResult.valid) {
         return authHeaderValidationResult.errorResponse;
@@ -41,65 +42,4 @@ export const getServiceClientsHandler = async (event: APIGatewayProxyEvent): Pro
         });
 
     return response;
-};
-
-const validateAuthorizationHeader = (
-    authHeader: string | undefined
-):
-    | {
-          valid: true;
-          userId: string;
-      }
-    | {
-          valid: false;
-          errorResponse: {
-              statusCode: number;
-              body: string;
-          };
-      } => {
-    if (!authHeader || !authHeader.startsWith("Bearer ")) {
-        return {
-            valid: false,
-            errorResponse: {
-                statusCode: 401,
-                body: "Missing access token"
-            }
-        };
-    }
-
-    const authToken = authHeader.substring(7);
-    //We trust the signature as this token is attached from
-    //the frontend which does the validation
-    const tokenParts = authToken.split(".");
-
-    if (tokenParts.length !== 3) {
-        return {
-            valid: false,
-            errorResponse: {
-                statusCode: 400,
-                body: "Invalid access token"
-            }
-        };
-    }
-
-    const encodedPayload = tokenParts[1];
-    const decodedPayload = Buffer.from(encodedPayload, "base64url").toString("utf-8");
-    const parsedPayload = JSON.parse(decodedPayload);
-
-    const subjectClaim = parsedPayload.sub;
-
-    if (typeof subjectClaim !== "string") {
-        return {
-            valid: false,
-            errorResponse: {
-                statusCode: 400,
-                body: "Invalid access token"
-            }
-        };
-    }
-
-    return {
-        valid: true,
-        userId: subjectClaim
-    };
 };
